app/requirements.txt
acres==0.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
audio-recorder-streamlit==0.0.10
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
ci-info==0.3.0
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
DateTime==5.5
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
docx2txt==0.8
durationpy==0.9
entrypoints==0.4
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.1.0
faiss-cpu==1.10.0
fastapi==0.115.8
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
gTTS==2.5.4
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.21
langchain-chroma==0.2.2
langchain-community==0.3.19
langchain-core==0.3.46
langchain-openai==0.3.9
langchain-pinecone==0.2.3
langchain-tests==0.3.14
langchain-text-splitters==0.3.7
langchain-upstage==0.6.0
langgraph==0.3.0
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.53
langsmith==0.3.11
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.67.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.17.8
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pluggy==1.5.0
posthog==3.16.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
prov==2.0.1
psutil==6.1.1
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.1
pykrx==1.0.48
PyMuPDF==1.25.3
pyparsing==3.2.1
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
streamlit==1.43.2
streamlit-audiorec==0.1.3
streamlit-audiorecorder==0.0.6
streamlit-chat==0.1.1
streamlit-option-menu==0.4.0
sympy==1.13.1
syrupy==4.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
tokenizers==0.19.1
toml==0.10.2
toolz==1.0.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.49.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
wrapt==1.17.2
xlrd==2.0.1
yarl==1.18.3
yfinance==0.2.52
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
